- Cut API calls in half during video upload. - Rename fix 1/2,
